+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-20T23:11:59.550Z",
+    "2026-06-21T08:27:58.801Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {